feat: work in factories to data in db and middleware role worked.

+++ app/Http/Controllers/CategoryController.php
@@ -38,7 +38,7 @@ class CategoryController extends Controller
         $image = $request->file('image_path');
         $nameImage = Str::uuid().'.'. $image->extension();
 
-        $imageServer = Image::make($image)->fit(200, 200);
+        $imageServer = Image::make($image)->fit(50, 50);
         $imagePath = public_path('categories').'/'.$nameImage;
         $imageServer->save($imagePath);
         $data['image_path'] = 'categories/'. $nameImage;
@@ -63,12 +63,15 @@ class CategoryController extends Controller
         $data = $request->all();
 
         if ($request->file('image_path')) {
+
             $image_path = public_path($category->image_path);
-            unlink($image_path);
+            if (file_exists($image_path)) {
+                unlink($image_path);
+            }
             $image = $request->file('image_path');
             $nameImage = Str::uuid().'.'. $image->extension();
 
-            $imageServer = Image::make($image)->fit(200, 200);
+            $imageServer = Image::make($image)->fit(50, 50);
             $imagePath = public_path('categories').'/'.$nameImage;
             $imageServer->save($imagePath);
             $data['image_path'] = 'categories/'. $nameImage;
